requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.0
opencv-python>=4.5.0
Pillow>=8.0.0
scipy>=1.6.0
tqdm>=4.60.0
wandb>=0.10.0
scikit-image>=0.18.0
imageio>=2.9.0

